third_party/WebKit/LayoutTests/virtual/enable-webauthn/http/tests/credentialmanager/credentialscontainer-create-from-nested-frame-expected.txt

CONSOLE WARNING: line 10: The following mojom is loaded multiple times: url/mojom/origin.mojom
CONSOLE WARNING: line 10: The following mojom is loaded multiple times: url/mojom/url.mojom
CONSOLE WARNING: line 10: The following mojom is loaded multiple times: mojo/public/mojom/base/time.mojom
CONSOLE WARNING: line 10: The following mojom is loaded multiple times: url/mojom/origin.mojom
CONSOLE WARNING: line 10: The following mojom is loaded multiple times: url/mojom/url.mojom
CONSOLE WARNING: line 10: The following mojom is loaded multiple times: mojo/public/mojom/base/time.mojom
This is a testharness.js-based test.
PASS navigator.credentials should be undefined in documents generated from `data:` URLs.
PASS navigator.credentials.create({publicKey}) in a javascript url should should succeed.
PASS navigator.credentials.create({publicKey}) in srcdoc should succeed.
PASS navigator.credentials.create({publicKey}) in about:blank embedded in a secure context should succeed.
PASS navigator.credentials.create({publicKey}) in an about:blank page embedded in a secure context should pass rpID checks.
Harness: the test ran to completion.
